Packet: initial encryption implementation
The concept is to use a symmetric key that is based on the password hash,
which is known to the client and server (but is only transferred upon
registration and password change).
This key and a plaintext number is used to seed a PRNG function to XOR
the packet data.

(This feature is yet unused)